refactor(matcha-window): split WindowConfig into native/web modules
`window_config/common.rs` holds the platform-agnostic value types
(`Size`, `Position`, `WindowButtons`, `Theme`, `Fullscreen`) and their
winit conversions. `window_config/native.rs` and
`window_config/web.rs` each define a `WindowConfig` struct with only
the fields and builders relevant to that target, plus
`to_winit_attributes` and a new `from_winit_window` constructor.

`winit_window.rs::get_config` now delegates to
`WindowConfig::from_winit_window`, removing the last cfg site in that
file.
